refactor: move stream stuff to its own module
